requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
deepagents
langchain
# langchain-openai==0.1.8
# openai==1.30.0
python-dotenv==1.0.0
langchain-anthropic
anthropic